Moved non-gui code for weather widget
Moved all non-gui code for weather widget into its own modules
Rewrote how some api works in the background
