+++ src/ngb/modules/weather_modules/__init__.py
@@ -1,0 +1,3 @@
+from .smhi import SMHI
+from .weatherbase import Weather_Base
+from .yr import YR
